% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_examples.R
\name{caret_examples}
\alias{caret_examples}
\title{Run All Examples of the 'caret' Package via Futureverse}
\usage{
caret_examples(exclude = NULL)
}
\arguments{
\item{exclude}{(optional) A character vector of Rd topics to exclude.}
}
\description{
Run All Examples of the 'caret' Package via Futureverse
}
