% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreach_examples.R
\name{foreach_examples}
\alias{foreach_examples}
\title{Run All Examples of the 'foreach' Package via Futureverse}
\usage{
foreach_examples()
}
\description{
Run All Examples of the 'foreach' Package via Futureverse
}
