% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_example}
\alias{run_example}
\title{Run one package example}
\usage{
run_example(
  topic,
  package,
  local = FALSE,
  run.dontrun = TRUE,
  envir = globalenv()
)
}
\arguments{
\item{topic}{name or literal character string: the online
    \code{\link[utils]{help}} topic the examples of which should be run.}

\item{package}{a character vector giving the package names to look
    into for the topic, or \code{NULL} (the default), when all packages on
    the \link{search} path are used.}

\item{local}{logical: if \code{TRUE} evaluate locally, if \code{FALSE}
    evaluate in the workspace.}

\item{run.dontrun}{logical indicating that \verb{\\dontrun}
    should be ignored.}

\item{envir}{The environment from which variables created by the
examples should be removed at the end.}
}
\value{
Invisibly, returns the processing time.
}
\description{
Run one package example
}
\seealso{
\code{\link[utils:example]{utils::example()}} is used to run the select example.
}
