% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{run_examples}
\alias{run_examples}
\title{Run a package's examples for a specific future plan}
\usage{
run_examples(package, topics = test_topics(package), strategy, ...)
}
\arguments{
\item{package}{A package name.}

\item{topics}{One or more Rd topics.}

\item{strategy}{A future strategy passed to \code{\link[future:plan]{future::plan()}}.}

\item{\dots}{Additional arguments passed to \code{\link[=run_example]{run_example()}}.}
}
\value{
Invisibly, returns the processing time.
}
\description{
Run a package's examples for a specific future plan
}
