% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_strategies}
\alias{test_strategies}
\title{Get Available Future Strategies}
\usage{
test_strategies()
}
\value{
A character vector of future strategies.
}
\description{
Get Available Future Strategies
}
