% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_topics}
\alias{test_topics}
\title{Run All Examples of the 'test_topics' Package via Futureverse}
\usage{
test_topics(package, subset = NA_integer_, max_subset = NULL)
}
\arguments{
\item{package}{A package name.}

\item{subset}{An integer index.}

\item{max_subset}{An integer or NULL.}
}
\value{
Return Rd topics.
}
\description{
Run All Examples of the 'test_topics' Package via Futureverse
}
